osmyaml/requirements.txt
pytest
click
requests
setuptools
pyyaml